Fix x64 version doesn't compile.
git-svn-id: https://subversion.assembla.com/svn/q3/trunk@3392 8af8166b-12a6-a448-a533-9086ace3f9f6

+++ q3/qm/src/ui/listwindow.cpp
@@ -1533,7 +1533,7 @@ LRESULT qm::ListWindow::onCreate(CREATESTRUCT* pCreateStruct)
 	};
 	Window toolTip(pImpl_->hwndToolTip_);
 	toolTip.sendMessage(TTM_ADDTOOL, 0, reinterpret_cast<LPARAM>(&ti));
-	int nInitialDelay = toolTip.sendMessage(TTM_GETDELAYTIME, TTDT_RESHOW);
+	int nInitialDelay = static_cast<int>(toolTip.sendMessage(TTM_GETDELAYTIME, TTDT_RESHOW));
 	toolTip.sendMessage(TTM_SETDELAYTIME, TTDT_INITIAL, MAKELONG(nInitialDelay, 0));
 #endif
